refactor(user): refactor supertokens recipes types
