+++ packages/user/src/index.ts
@@ -96,8 +96,8 @@ export { default as validatePassword } from "./validator/password";
 
 export * from "./constants";
 
-export type { EmailVerificationRecipe } from "./supertokens/emailVerificationRecipeConfigTypes";
-export type { ThirdPartyEmailPasswordRecipe } from "./supertokens/types";
+export type { EmailVerificationRecipe } from "./supertokens/types/emailVerificationRecipe";
+export type { ThirdPartyEmailPasswordRecipe } from "./supertokens/types/thirdPartyEmailPasswordRecipe";
 export type {
   AuthUser,
   ChangePasswordInput,
